Moved CcsRecordMonthlyAggregationHelper to own file
